Switch from "around" semantic to "lambdize"
Lambdize is much closer, semantically, to the behavior of what that API is doing. These are modeled after LambdaObserver/Subscriber, and I think it makes sense to describe it as such. It's a word used in other languages to describe this as well, and "around" was always a little awkward.