Fix ErrorsRule to fix master tests
RxJava 2.0.6 introduced UndeliverableException, which wraps everything that goes to RxJavaPlugins.onError() that isn't an NPE (to better distinguish between different types of exceptions).

Also opportunistically fixed a deprecation from the bump to Truth 0.32

This was missed because the lambda observer tests went into master and then the deps bump went in after, but we don't have CI turned on to check that the tests would have failed in between those two changes.