Adjust backfilling to handle state->event corresponding events correctly
This resolves #93

I misunderstood how the state->event relationship worked, and the event always precedes the state. The result is that if we want to do the mapping in a way that properly handles corresponding events, we need to shift everything up by 1. Now, initialized means we'll mark its event as "ON_CREATE", and started/resumed both point to ON_RESUME (because started is still the state after onResume and resumed is marked after that callback)