Migrate ObservableScoper to delegate to the as() APIs
This migrates ObservableScoper to delegate to the as() APIs. This was a bit weird because the scoper APIs were a bit more flexible (due to the nature of to()) with generics, and we have to cover that with an identity function. This makes them slightly more overhead now, but we're ok with that because they're a deprecated API that we're encouraging people to migrate off of.