Refactor and simplify ViewScopeProvider to just implement ScopeProvider
This refactors ViewScopeProvider to just implement ScopeProvider, as it should really. It only ever considered one boundary (detached), and was a lot of overhead for it. This slims it down to just implement ScopeProvider and throw the boundary check manually.

This also reduces overhead by not eagerly subscribing to lifecycle events on the view and only creating a Maybe on demand of the detach event.

Finally - this removes the previously erroneously-exposed `ViewLifecycleEvent` enum, which was never supposed to be public.