Switch to Completable for scope provisioning
This is a significant API change, but a good one we want to clean up before releasing. `Maybe` seemed like the right idea for something that "may or may not emit", but in our case we actually don't care about the difference between onSuccess or onComplete. We did have a notion of "UNBOUND", but that doesn't offer anything other than a severed lifecycle scope disposal in an atomic reference (no other cleanups would happen for gc, etc). This brings us to a `Single`. The thing is though, we don't care about the object/element type. A `Single` where the type doesn't matter is semantically a `Completable`, and thus this change.